+++ R/coerce-methods.R
@@ -21,3 +21,10 @@ setAs(from = "lib_df", to = "data.frame", def = function(from) {
 setAs(from = "lib_df", to = "bibentry", def = function(from) {
   return(bib2bibentry(from))
 })
+
+#' @name as
+#' @rdname coerce-methods
+#' @aliases coerce,bibentry,lib_df-method
+setAs(from = "bibentry", to = "lib_df", def = function(from) {
+  return(read_bib(capture.output(print(from, "bibtex"))))
+})