+++ R/imports.R
@@ -2,7 +2,7 @@
 #' @importFrom stats na.omit aggregate
 #' @importFrom stringr fixed regex str_count str_match_all str_trim
 #' @importFrom tools file_path_sans_ext
-#' @importFrom utils bibentry
+#' @importFrom utils bibentry capture.output
 #' @importFrom yamlme render_rmd write_rmd
 #' @importClassesFrom yamlme rmd_doc
 NULL