+++ lab/exp-read_bib.R
@@ -19,17 +19,6 @@ Bib
 as(Bib, "data.frame")
 
 # Solving issue #49
-
 Refs <- read_bib(x = "lab/data/MiguelReferences.bib")
-Keys <- readLines("lab/data/Keys.txt")
-Refs <- subset(Refs, bibtexkey %in% Keys)
-write_bib(Refs, "lab/data/references")
-
-as(Refs, "data.frame")
-
 x <- subset(Refs, grepl("address =", address))
 as(x, "data.frame")
-
-
-
-